+++ lib/application_factory_static.dart
@@ -26,7 +26,6 @@ library angular.app.factory.static;
 
 import 'package:angular/application.dart';
 import 'package:angular/core/registry.dart';
-import 'package:angular/core/parser/dynamic_parser.dart';
 import 'package:angular/core/parser/parser.dart';
 import 'package:angular/core/parser/static_closure_map.dart';
 import 'package:angular/core/registry_static.dart';
