+++ test/angular_spec.dart
@@ -147,7 +147,7 @@ main() {
         "angular.core.dom_internal.ViewFactory",
         "angular.core.dom_internal.ViewPort",
         "angular.core.parser.Parser",
-        "angular.core.parser.dynamic_parser.ClosureMap",
+        "angular.core.parser.ClosureMap",
         "angular.core_internal.ExceptionHandler",
         "angular.core_internal.Interpolate",
         "angular.core_internal.RootScope",
@@ -159,6 +159,8 @@ main() {
         "angular.core_internal.ScopeStatsConfig",
         "angular.core_internal.ScopeStatsEmitter",
         "angular.core_internal.VmTurnZone",
+        "angular.core.parser.ClosureMap",
+        "angular.core.parser.Parser",
         "angular.directive.AHref",
         "angular.directive.ContentEditable",
         "angular.directive.DirectiveModule",
