+++ test/core/parser/parser_spec.dart
@@ -57,17 +57,6 @@ main() {
       module.bind(SubstringFormatter);
     });
 
-    describe('DynamicParser', () {
-      // This is important because the DynamicParser expects to be a singleton
-      // to share its cache.  It therefore registers with the CacheRegister and
-      // having more than one instance will result in a duplicate registration.
-      it('should be identical to Parser in dynamic mode', (Parser p, DynamicParser dp) {
-        if (p is DynamicParser) {
-          expect(identical(p, dp)).toBeTruthy();;
-        }
-      });
-    });
-
     beforeEach((Parser injectedParser, FormatterMap injectedFormatters) {
       parser = injectedParser;
       formatters = injectedFormatters;
@@ -173,12 +162,6 @@ main() {
       beforeEach((Parser p) {
         parser = p;
       });
-
-      // We only care about the error strings in the DynamicParser.
-      var errStr = (x) {
-        if (parser is DynamicParser) { return x; }
-        return null;
-      };
 
       // PARSER ERRORS
       it('should throw a reasonable error for unconsumed tokens', () {
